szkielet aplikacji

+++ src/main/java/agh/ics/oop/model/MapOptions.java
@@ -1,4 +1,4 @@
 package agh.ics.oop.model;
 
-public record MapOptions(int mapHeight, int mapWidth, int startingNumOfPlants, int plantNumEveryDay, int startingNumOfAnimals, double energyStart) {
+public record MapOptions(int mapHeight, int mapWidth, int startingNumOfPlants, int plantNumEveryDay, int startingNumOfAnimals, int energyStart) {
 }
